Swath Profile

Calculate swath-profile values perpendicular to a straight baseline. The baseline is generated between two user-defined points (X|Y), see argument coords. The distance between samples and the number of samples can be specified, see arguments k and dist. Values of the swath-profile are extracted from a given raster file, see argument raster. CRS of raster and points have to be the same.

Usage

swath_extract(
  coords,
  raster,
  k = 1,
  dist,
  crs = "EPSG:4326",
  method = c("bilinear", "simple")
)

Source

The algorithm is a modified version of "swathR" by Vincent Haburaj (https://github.com/jjvhab/swathR).

Arguments

coords

either a sf object or a matrix(ncol=2, nrow=2) with x and y coordinates of beginning and end point of the baseline; each point in one row

column 1

x coordinates

column 2

y coordinates

raster

Raster file (loaded with terra::rast())

k

integer. number of lines on each side of the baseline

dist

numeric. distance between lines

crs

character. coordinate reference system.

method

character. method for extraction of raw data, see terra::extract(): default value: "bilinear"

Value

list.

swath

matrix. Statistics of the raster measured along the lines

data

list of numeric vector containing the data extracted from the raster along each line

lines

list of of the lines as SpatVectors

See also

Examples

# Create a random raster
r <- terra::rast(ncol = 10, nrow = 10, xmin = -150, xmax = -80, ymin = 20, ymax = 60)
values(r) <- runif(terra::ncell(r))
#> Error in values(r) <- runif(terra::ncell(r)): could not find function "values<-"

# Create a random profile
profile <- data.frame(lon = c(-140, -90), lat = c(55, 25)) |>
  sf::st_as_sf(coords = c("lon", "lat"), crs = "WGS84")
swath_extract(profile, r, k = 2, dist = 1)
#> Error: [extract] raster has no values